Change protocol for sensory phases and refactor
In sensory phases, not a single random stimulus is shown for the entire
duration, but the stimulus should change randomly at specified
intervals.
The chosen solution as of now is to reflect this in the protocol, i.e.
give instruction for a new random stimulus at specified times with the
given interval. As the interval is typically chosen to be very small,
this results in a lot of instructions. If it turns out to be a problem
for the simulation Stimulator class to hold them all in memory, can
either consider to read the protocol chunk-wise, or implement the
randomisation of stimulus during sensory phases into the Stimulator
itself.